requirements.txt
requests>=2.31.0
numpy>=1.26.0
streamlit>=1.35.0
plotly>=5.19.0
fpdf2>=2.7.9
